+++ 17-plugin-pmd/00.pmd/build.gradle
@@ -1,0 +1,25 @@
+apply plugin: 'java'
+apply plugin: 'pmd'
+
+ext {
+	p3c = "1.3.0"
+}
+
+pmd {
+	consoleOutput = true
+    reportsDir = file("build/reports/pmd")
+    ignoreFailures = true
+
+	ruleSets = [
+		"java-ali-comment",
+		"java-basic"
+	]
+}
+
+repositories {
+   jcenter()
+}
+
+dependencies {
+	pmd "com.alibaba.p3c:p3c-pmd:${p3c}"
+}
