+++ 17-plugin-pmd/01.alipmd/src/main/java/Tiger.java
@@ -1,0 +1,20 @@
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * Tiger say hi
+ */
+public class Tiger {
+    public static void sayHi(){
+        System.out.println("Hi~");
+    }
+
+	public void hmap(){
+		Map<String,String> mp = new HashMap<String,String>();//新建map
+		//int a=1
+		//int b=2
+	}
+    public static void main(String[] args) {
+        System.out.println("PKAQ");
+    }
+}
