--- 17-plugin-pmd/build.gradle
@@ -1,23 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'pmd'
-
-ext {
-	p3c = "1.3.0"
-}
-
-pmd {
-	consoleOutput = true
-    reportsDir = file("build/reports/pmd")
-
-	ruleSets = [
-		"java-ali-comment"
-	]
-}
-
-repositories {
-   jcenter()
-}
-
-dependencies {
-	pmd "com.alibaba.p3c:p3c-pmd:${p3c}"
-}
