+++ 18-incremental-artifact/README.MD
@@ -1,0 +1,22 @@
+
+### 打包src/main/java中的资源文件
+
+可能有的同学会有疑问，为什么会有此篇。
+实际上这个需求在老项目和采用mybatis的项目上体现的比较突出。
+但作为一线程序员往往因不可抗力因素是无法凭借一己之力将资源文件随便就挪动位置的。
+so...
+
+解决此问题思路有几种  
+1.移动资源文件到resources下  
+2.将资源文件包含进来  
+3.将resources指向目标目录  
+
+
+本文档采用的是第二种，至于第三种有兴趣的同学可以更改sourceset进行操作  
+特别说明:
+由于gradle在build时会把类文件和资源文件分别放到 build/classes,build/resources下
+所以classes下仍然是无法看到资源文件的,但不影响打包
+如示例所示,修改了jar任务的一些行为,所以会在打jar的时候生效
+如果是要打war包,稍微拓展一下 你就知道啦
+
+p.s 关于此篇文件夹名称头两个单词 让我想起了 LOST
