+++ 18-incremental-artifact/build.gradle
@@ -1,0 +1,19 @@
+apply plugin: 'java'
+compileJava {
+    options.incremental = true
+    options.listFiles = true
+    logging.addStandardOutputListener(new LogListener(project))
+}
+
+class LogListener implements org.gradle.api.logging.StandardOutputListener {
+    
+    private Project project
+    
+    public LogListener(Project project) {
+        this.project = project
+    }
+
+    public void onOutput(CharSequence output) {
+       println this.project
+    }
+}
