+++ 18-incremental-artifact/build.gradle.bak
@@ -1,0 +1,64 @@
+apply plugin: 'java'
+
+
+jar {
+    from('src/main/java') {
+   	 include 'config/**/*.xml'
+   }
+
+   from('src/main/java') {
+   	 include 'prop/**/*.properties'
+   }
+
+}
+
+compileJava {   
+    options.incremental = true
+ 	
+}
+
+
+
+task incrementalReverse(type: IncrementalReverseTask) {  
+    inputDir = file("$projectDir/src/main/java")
+    outputDir = file("$buildDir/outputs")
+    inputProperty = project.properties['taskInputProperty'] ?: "original"
+
+     // 存在更新      
+    doLast {
+        if(outputs.getHasOutput()){
+            println  outputs.getFiles().getAsFileTree().getFiles()
+        }
+    }
+}
+
+
+class IncrementalReverseTask extends DefaultTask {
+    @InputDirectory
+    def File inputDir
+
+    @OutputDirectory
+    def File outputDir
+
+    @Input
+    def inputProperty
+
+    @TaskAction
+    void execute(IncrementalTaskInputs inputs) {
+        println inputs.incremental ? "CHANGED inputs considered out of date"
+                                   : "ALL inputs considered out of date"
+        if (!inputs.incremental)
+            project.delete(outputDir.listFiles())
+        // 存在更新 获取变更文件的文件名
+        inputs.outOfDate { change ->
+            println "out of date: ${change.file.name}"
+            def targetFile = new File(outputDir, change.file.name)
+            targetFile.text = change.file.text.reverse()
+        }        
+        // 执行编译
+
+        // 抽取增量结果
+
+        // 打zip包
+    }
+}
