+++ 18-incremental-artifact/src/main/java/Tiger.java
@@ -1,0 +1,8 @@
+/**
+ * Tiger say hi
+ */
+public class Tiger {
+    public static void sayHi(){
+        System.out.println("c11111111111~");
+    }
+}
