+++ 18-incremental-artifact/src/main/java/Tx.java
@@ -1,0 +1,8 @@
+/**
+ * Tiger say hi
+ */
+public class Tx {
+    public static void sayHi(){
+        System.out.println("Hi~");
+    } 
+}
