+++ 19/build.gradle
@@ -1,0 +1,11 @@
+apply plugin: 'java'
+task dotDescription(type: JavaExec, dependsOn: "classes"){
+   classpath configurations.runtime, sourceSets.main.output
+   main = "Tiger"
+   
+   standardOutput = new ByteArrayOutputStream()
+	doLast {
+	  def theOutput = standardOutput.toString()
+	  println theOutput
+	}
+}
