+++ 19/src/main/java/Tiger.java
@@ -1,0 +1,12 @@
+/**
+ * Tiger say hi
+ */
+public class Tiger {
+    public static void sayHi(){
+        System.out.println("Hi~");
+    }
+
+    public static void main(String[] args){
+		System.out.println("H1i~");
+    }
+}
